chore: add eslint and fix components names
