+++ src/Details/styled.js
@@ -1,14 +1,14 @@
-import styled from "styled-components"
-import { media } from "../styled-components/media"
+import styled from 'styled-components';
+import media from '../styled-components/media';
 
 export const Wrapper = styled.section`
   padding: 1.1rem 0;
   flex-basis: 100%;
-`
+`;
 
 export const Content = styled.div`
   padding: 1rem 0;
-`
+`;
 
 export const BuyNowBtn = styled.button`
   margin: 0;
@@ -26,25 +26,25 @@ export const BuyNowBtn = styled.button`
   text-align: center;
   color: #fff;
   background: linear-gradient(to right, #4949aa, #27275d);
-  ${media.xs("min")`
+  ${media.xs('min')`
     margin-left: 34.5%;
     padding: 2rem 0;
     width: 65.5%;
   `}
-  ${media.lg("min")`
+  ${media.lg('min')`
     margin-left: 414px;
     padding: 3rem 0;
     width: calc(100% - 414px);
   `}
-`
+`;
 export const Header = styled.div`
   margin-bottom: 3rem;
-`
+`;
 export const ButtonsWrapper = styled.div`
   display: flex;
   align-items: center;
   justify-content: flex-end;
-`
+`;
 
 export const ButtonColor = styled.button`
   display: inline-block;
@@ -54,7 +54,7 @@ export const ButtonColor = styled.button`
   height: 18px;
   border-radius: 9px;
   background: ${props => props.color};
-`
+`;
 
 export const Title = styled.h1`
   margin: 0;
@@ -64,7 +64,7 @@ export const Title = styled.h1`
   line-height: 1;
   color: #3a3a3a;
   text-transform: uppercase;
-`
+`;
 
 export const Price = styled.h2`
   margin: 0;
@@ -73,7 +73,7 @@ export const Price = styled.h2`
   font-weight: 700;
   font-size: 80px;
   color: #e2e2e2;
-`
+`;
 
 export const ButtonSave = styled.button`
   width: 75px;
@@ -86,11 +86,11 @@ export const ButtonSave = styled.button`
   background: #e3e3e3;
   text-transform: uppercase;
   border-radius: 38px;
-`
+`;
 
 export const PhotosWrapper = styled.div`
   margin-bottom: 3rem;
-`
+`;
 
 export const Lines = styled.span`
   position: relative;
@@ -116,14 +116,15 @@ export const Lines = styled.span`
   &:after {
     bottom: -5px;
   }
-`
+`;
 
 export const ButtonMore = styled.div`
   font-family: "avenir";
   font-size: 14px;
   text-align: left;
   color: #c0c0c0;
-`
+`;
+
 export const Description = styled.p`
   margin-bottom: 15rem;
   font-family: "avenir";
@@ -136,4 +137,4 @@ export const Description = styled.p`
     color: #6e6e6e;
     font-weight: 400;
   }
-`
+`;
