--- src/Sidebar/Nav/styles.css
@@ -1,3 +1,0 @@
-.menu {
-  text-align: center;
-}
