--- src/Sidebar/Search/index.js
@@ -1,10 +1,0 @@
-import React from "react"
-import { Form, Input } from "./styled"
-
-export default () => {
-  return (
-    <Form isSearch>
-      <Input />
-    </Form>
-  )
-}
