+++ src/Sidebar/Search/index.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Form, Input } from './styled';
+
+export default () => (
+  <Form isSearch>
+    <Input />
+  </Form>
+);
