+++ src/Sidebar/index.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Logo from './Logo';
+import Search from './Search';
+import Nav from './Nav';
+import { Wrapper, HideMobile } from './styled';
+
+export default () => (
+  <Wrapper>
+    <Logo />
+    <HideMobile>
+      <Search />
+      <Nav />
+    </HideMobile>
+  </Wrapper>
+);
