+++ src/index.js
@@ -1,5 +1,4 @@
 /* eslint-disable react/jsx-filename-extension */
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
