--- src/styled-components/config.js
@@ -1,8 +1,0 @@
-export default {
-  media: {
-    xs: "468px",
-    sm: "768px",
-    md: "970px",
-    lg: "1439px",
-  },
-}
