+++ src/styled-components/config.jsx
@@ -1,0 +1,8 @@
+export default {
+  media: {
+    xs: '468px',
+    sm: '768px',
+    md: '970px',
+    lg: '1439px',
+  },
+};
